24.1110 FIX -> Conversión de campo -id- del usuario a String        Pasaba que no dejaba traiga la información ni dejaba eliminar la cuenta en profile        Modificaciones en interfaz de user y en el register ADD -> Posibilidad al usuario de volver hacia atrás en profile        Reutilicé el componente -Navbar- en dicho componente        Utilicé ngIf para no mostrar el search y el ícono de notificaciones. También agregué botón para volver al home. MOD -> Lógica en la función -confirmDelete-        Agregué la validación de que sí el usuario no ingresa un email y desea cancelar la operación, no siga la ejecución de la función.        Esto es útil para que no aparezca el alert de -Email incorrecto- sí el usuario toca en cancelar.

+++ src/app/pages/profile-page/profile-page.component.ts
@@ -49,6 +49,9 @@ export class ProfilePageComponent implements OnInit {
 
   confirmDelete(): void {
     const email = prompt('Confirma tu email para eliminar la cuenta:');
+
+    if (!email) return;
+    
     if (email === this.user?.email) {
       this.http.delete(`http://localhost:3000/users/${this.user.id}`).subscribe(
         () => console.log('User deleted successfully'),
